move AssignPermissionsToDataSet to permissions directory
