+++ src/containers/org/components/index.js
@@ -2,8 +2,6 @@
  * @flow
  */
 
-import { AssignPermissionsToDataSet } from './assign-permissions-to-data-set';
-
 export { default as AddRoleToOrgModal } from './AddRoleToOrgModal';
 export { default as CreateOrgModal } from './CreateOrgModal';
 export { default as DataSetSearchResultCard } from './DataSetSearchResultCard';
@@ -13,7 +11,3 @@ export { default as RemoveRoleFromMemberModal } from './RemoveRoleFromMemberModa
 export { default as RenameOrgDatabaseModal } from './RenameOrgDatabaseModal';
 export { default as RoleActionButton } from './RoleActionButton';
 export { default as RoleDetailsModal } from './RoleDetailsModal';
-
-export {
-  AssignPermissionsToDataSet,
-};
