+++ src/containers/permissions/index.js
@@ -2,7 +2,13 @@
  * @flow
  */
 
+import { AssignPermissionsToDataSet } from './assign-permissions-to-data-set';
+
 export { default as DataSetPermissionsContainer } from './DataSetPermissionsContainer';
 export { default as ObjectPermissionsContainer } from './ObjectPermissionsContainer';
 export { default as PermissionsActionsGrid } from './PermissionsActionsGrid';
 export { default as PermissionsPanel } from './PermissionsPanel';
+
+export {
+  AssignPermissionsToDataSet,
+};
